[libc++] Implement LWG4169 constrain `std::atomic<T>`'s default constructor
Drive-by: Rename `constexpr_noexcept.compile.pass.cpp` to
`ctor.default.pass.cpp` as test coverage is added to it, and compile and
run it in all modes.
